pyproject.toml
[project]
name = "product-launch-war-room"
version = "0.1.0"
description = "Multi-agent system for product launch decision support"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-groq>=0.2.0",
    # Removed: crewai, crewai-tools
    "llama-index>=0.12.0",
    "llama-index-embeddings-huggingface>=0.4.0",
    "pydantic>=2.9.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "rich>=13.9.0",
    "langsmith>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "black>=24.8.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88